:truck: chore(api) move obfuscation service to `src/shared`
