+++ api/src/shared/domain/services/obfuscation-service.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../src/identity-access-management/domain/constants/identity-providers.js';
-import * as authenticationMethodRepository from '../../../src/identity-access-management/infrastructure/repositories/authentication-method.repository.js';
-import { NotFoundError } from '../../../src/shared/domain/errors.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../identity-access-management/domain/constants/identity-providers.js';
+import * as authenticationMethodRepository from '../../../identity-access-management/infrastructure/repositories/authentication-method.repository.js';
+import { NotFoundError } from '../../domain/errors.js';
 
 const CONNEXION_TYPES = {
   username: 'username',
