+++ api/tests/shared/unit/domain/services/obfuscation-service_test.js
@@ -1,8 +1,8 @@
-import * as obfuscationService from '../../../../lib/domain/services/obfuscation-service.js';
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../src/identity-access-management/domain/constants/identity-providers.js';
-import { User } from '../../../../src/identity-access-management/domain/models/User.js';
-import { NotFoundError } from '../../../../src/shared/domain/errors.js';
-import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../../src/identity-access-management/domain/constants/identity-providers.js';
+import { User } from '../../../../../src/identity-access-management/domain/models/User.js';
+import { NotFoundError } from '../../../../../src/shared/domain/errors.js';
+import * as obfuscationService from '../../../../../src/shared/domain/services/obfuscation-service.js';
+import { catchErr, domainBuilder, expect, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Service | user-authentication-method-obfuscation-service', function () {
   let authenticationMethodRepository;
